requirements.txt
google-generativeai==0.8.4
pillow==10.2.0
streamlit==1.44.0
streamlit_option_menu==0.4.0